+++ planet/api/models.py
@@ -14,6 +14,7 @@
 """Manage data for requests and responses."""
 import copy
 import datetime
+import json
 import logging
 
 import httpx
@@ -101,83 +102,6 @@ class Response():
         await self.http_response.aclose()
 
 
-class Paged():
-    LINKS_KEY = 'links'
-    NEXT_KEY = 'next'
-    ITEMS_KEY = 'items'
-
-    def __init__(self, request, do_request_fcn, limit=None):
-        self.request = request
-        self._do_request = do_request_fcn
-
-        self._pages = None
-        self._items = []
-
-        self.i = 0
-        self.limit = limit
-
-    def __aiter__(self):
-        return self
-
-    async def __anext__(self):
-        # This was implemented because traversing _get_pages()
-        # in an async generator was resulting in retrieving all the
-        # pages, when the goal is to stop retrieval when the limit
-        # is reached
-        if self.limit is not None and self.i >= self.limit:
-            raise StopAsyncIteration
-
-        try:
-            item = self._items.pop(0)
-            self.i += 1
-        except IndexError:
-            self._pages = self._pages or self._get_pages()
-            page = await self._pages.__anext__()
-            self._items = page[self.ITEMS_KEY]
-            try:
-                item = self._items.pop(0)
-                self.i += 1
-            except IndexError:
-                raise StopAsyncIteration
-
-        return item
-
-    # async def aiter_items(self, limit=None):
-    #     # why o why does this retrieve all the pages instead of stopping
-    #     # when count has reached its limit?
-    #     count = 0
-    #     async for page in self._get_pages():
-    #         for item in self._page_items(page):
-    #             count += 1
-    #             if not limit or count <= limit:
-    #                 yield item
-
-    async def _get_pages(self):
-        request = copy.deepcopy(self.request)
-        LOGGER.debug('getting first page')
-        resp = await self._do_request(request)
-        page = resp.json()
-        yield page
-
-        next_url = self._next_link(page)
-        while(next_url):
-            LOGGER.debug('getting next page')
-            request.url = next_url
-            resp = await self._do_request(request)
-            page = resp.json()
-            yield page
-            next_url = self._next_link(page)
-
-    def _next_link(self, page):
-        try:
-            next_link = page[self.LINKS_KEY][self.NEXT_KEY]
-            LOGGER.debug(f'next: {next_link}')
-        except KeyError:
-            LOGGER.debug('end of the pages')
-            next_link = False
-        return next_link
-
-
 class StreamingBody():
     '''A representation of a streaming resource from the API.
 
@@ -256,3 +180,135 @@ class StreamingBody():
                         previous = new
         except FileExistsError:
             LOGGER.info(f'File {filename} exists, not overwriting')
+
+
+class Paged():
+    LINKS_KEY = 'links'
+    NEXT_KEY = 'next'
+    ITEMS_KEY = 'items'
+    TYPE = None
+
+    def __init__(self, request, do_request_fcn, limit=None):
+        self.request = request
+        self._do_request = do_request_fcn
+
+        self._pages = None
+        self._items = []
+
+        self.i = 0
+        self.limit = limit
+
+    def __aiter__(self):
+        return self
+
+    async def __anext__(self):
+        # This was implemented because traversing _get_pages()
+        # in an async generator was resulting in retrieving all the
+        # pages, when the goal is to stop retrieval when the limit
+        # is reached
+        if self.limit is not None and self.i >= self.limit:
+            raise StopAsyncIteration
+
+        try:
+            item = self._items.pop(0)
+            self.i += 1
+        except IndexError:
+            self._pages = self._pages or self._get_pages()
+            page = await self._pages.__anext__()
+            self._items = page[self.ITEMS_KEY]
+            try:
+                item = self._items.pop(0)
+                self.i += 1
+            except IndexError:
+                raise StopAsyncIteration
+
+        return item
+
+    async def _get_pages(self):
+        request = copy.deepcopy(self.request)
+        LOGGER.debug('getting first page')
+        resp = await self._do_request(request)
+        page = resp.json()
+        yield page
+
+        next_url = self._next_link(page)
+        while(next_url):
+            LOGGER.debug('getting next page')
+            request.url = next_url
+            resp = await self._do_request(request)
+            page = resp.json()
+            yield page
+            next_url = self._next_link(page)
+
+    def _next_link(self, page):
+        try:
+            next_link = page[self.LINKS_KEY][self.NEXT_KEY]
+            LOGGER.debug(f'next: {next_link}')
+        except KeyError:
+            LOGGER.debug('end of the pages')
+            next_link = False
+        return next_link
+
+
+class Order():
+    '''Managing description of an order returned from Orders API.
+
+    :param data: Response json describing order
+    :type data: dict
+    '''
+    LINKS_KEY = '_links'
+    RESULTS_KEY = 'results'
+    LOCATION_KEY = 'location'
+
+    def __init__(self, data):
+        self.data = data
+
+    def __str__(self):
+        return "<Order> " + json.dumps(self.data)
+
+    @property
+    def results(self):
+        '''Results for each item in order.
+
+        :return: result for each item in order
+        :rtype: list of dict
+        '''
+        links = self.data[self.LINKS_KEY]
+        results = links.get(self.RESULTS_KEY, None)
+        return results
+
+    @property
+    def locations(self):
+        '''Download locations for order results.
+
+        :return: download locations in order
+        :rtype: list of str
+        '''
+        return list(r[self.LOCATION_KEY] for r in self.results)
+
+    @property
+    def state(self):
+        '''State of the order.
+
+        :return: state of order
+        :rtype: str
+        '''
+        return self.data['state']
+
+    @property
+    def id(self):
+        '''ID of the order.
+
+        :return: id of order
+        :rtype: str
+        '''
+        return self.data['id']
+
+
+class OrdersPaged(Paged):
+    LINKS_KEY = '_links'
+    NEXT_KEY = 'next'
+    ITEMS_KEY = 'orders'
+
+    async def __anext__(self):
+        return Order(await super().__anext__())
