+++ requirements.txt
@@ -1,3 +1,2 @@
-aiostream==0.4.1
 httpx==0.16.1
 tqdm==4.56.0
